requirements.txt
asgiref==3.11.0
asttokens==2.4.1
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.1.0
fonttools==4.61.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
joblib==1.5.3
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.4.2
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049096285/work
pandas==3.0.0
parso==0.8.4
pillow==12.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==26.2.0
reportlab==4.4.9
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.2.13
